requirements.txt
fastapi>=0.111.0
pydantic>=2.5
uvicorn[standard]==0.24.0
requests

sqlalchemy==2.0.28
psycopg2-binary

passlib[bcrypt]==1.7.4
bcrypt==4.0.1

python-jose[cryptography]==3.3.0
email-validator

httpx==0.25.2

pytest
alembic
python-multipart

y-py